refactor: Add side menu into its own component
